Alternative approach for findSchemaChanges (#4235)
Alternative to https://github.com/graphql/graphql-js/pull/2181

This adds a new `findSchemaChanges` export and deprecates the breaking
and dangerous alternative to be removed in v18. While adding
safe-schema-changes I noticed that we lack directive arg validation
here, we should probably add these to breaking/dangerous changes in the
future.